Added "loci.workflow" package.
Instead of having a singleton that passes images between plugins, this provides a workflow class.  A workflow is built up of components, the components have named inputs and outputs which are chained together.  The workflow is itself a component, so it also has named inputs and outputs.

The component and workflow can be saved to and restored from XML files (implementation uses a limited subset of simple XML).  The XML for a workflow also includes the XML for its containers, which may be other workflows.